requirements.txt
ansicolors==1.1.8
docker==2.6.1
h5py==2.7.1
matplotlib==2.1.0
numpy==1.13.3
pyyaml==3.12
psutil==5.4.2
scipy==1.0.0
scikit-learn==0.19.1
jinja2==2.10
kafka-python
